move find_first to patterns
