+++ noodles/__init__.py
@@ -1,7 +1,7 @@
 from noodles.interface import (
     delay, gather, lift, schedule, schedule_hint, unwrap,
     has_scheduled_methods, update_hints, unpack, quote, unquote,
-    find_first, gather_dict, result, gather_all, maybe, Fail)
+    gather_dict, result, gather_all, maybe, Fail)
 from .patterns import (filter, fold)
 from .run.runners import run_parallel_with_display as run_logging
 from .run.runners import (run_single, run_parallel)
@@ -16,4 +16,4 @@ __all__ = ['schedule', 'schedule_hint', 'run_single', 'run_process',
            'Scheduler', 'Storable', 'has_scheduled_methods', 'Fail',
            'fold', 'filter',
            'gather', 'gather_all', 'gather_dict', 'lift', 'unpack', 'maybe',
-           'delay', 'update_hints', 'quote', 'unquote', 'find_first', 'result']
+           'delay', 'update_hints', 'quote', 'unquote', 'result']
