+++ noodles/interface/__init__.py
@@ -3,7 +3,7 @@ from .decorator import (
     unwrap, update_hints, result)
 from .functions import (
     delay, gather, gather_all, gather_dict, lift, unpack, quote,
-    unquote, find_first)
+    unquote)
 from .annotated_value import (AnnotatedValue)
 from .exceptions import (JobException)
 from .maybe import (maybe, Fail)
@@ -11,5 +11,5 @@ from .maybe import (maybe, Fail)
 __all__ = ['delay', 'gather', 'gather_all', 'gather_dict', 'schedule_hint',
            'schedule', 'unpack', 'has_scheduled_methods', 'unwrap',
            'update_hints', 'AnnotatedValue', 'JobException', 'lift',
-           'PromisedObject', 'quote', 'unquote', 'find_first', 'result',
+           'PromisedObject', 'quote', 'unquote', 'result',
            'maybe', 'Fail']
