+++ noodles/patterns/__init__.py
@@ -1,3 +1,4 @@
 from .functional_patterns import (filter, fold)
+from .find_first import (find_first)
 
-__all__ = ['filter', 'fold']
+__all__ = ['filter', 'fold', 'find_first']
